requirements.txt
# Core dependencies
comet-ml>=3.35.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6

# Optional: Open source models
# transformers>=4.36.0
# torch>=2.1.0
